Add database configuration for Heroku and create library app
This commit adds new settings and code that allow the application to run on Heroku, and create the library app. Now the Django application can use either SQLite locally or the Postgres provided by Heroku in production. The library app includes models for Book and Author, views, tests, and demo data command, reflecting the necessary migrations. Several auxiliary files have been added or updated for Heroku compatibility and correct functioning of the app.